+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Upcoming
+
+* Add `tempFiles` helper
+
 ### 4.2.0
 
 * Use `consistent-path` package to determine `$PATH` correctly on OSX